requirements.txt
bs4==0.0.2
Flask==3.0.2
lxml==5.1.0
python-dotenv==1.0.1
requests @ file:///croot/requests_1690400202158/work
